Add files via upload
This was my first time making use of pandas for data cleaning and it will interest you to know that i found it very thrilling.
This data was imported into pandas as a csv file and had several issues; some of which i would mention.
1. There were about a 100 duplicate records which i had to delete from the data
2. There were several empty cells lacking data. I rectified this using the backward and forward fill
3. The dates were inconsistent but wit the help of pandas datetime i was able to maintain a consistent format across the column.
4.  There were typographical errors which i corrected and payment name were standardised
5. Some of the values denoting quantity and age were recorded as negative values. These negative was stripped of the values.

Finally, there were other problems with this data which i identified, however, i didn't make any changes to them to avoid tampering with the integrity of the data. The total sales is supposed to be a multiplication of the price and quantity of products sold. However, the values recorded were not the expected values after calculation. Also, some of the ages of the customers happened to be 1 or 2 years which is not realistic.